✨ 기능 추가: RecommendedWorkoutHistoryController 만족도 등록 controller 추가

+++ healthtart/src/main/java/com/dev5ops/healthtart/recommended_workout_history/domain/vo/request/RequestRegisterRecommendedWorkoutHistoryVO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class RequestRegisterRecommendedWorkoutHistoryVO {
-    private int routineRatings;
+    private Double routineRatings;
     private Long routineCode;
 }
